+++ slacky/__main__.py
@@ -8,7 +8,8 @@ import re
 print(Prefixes.event + 'Loading Plugins')
 
 commands = {
-        'heartbeat': lambda **payload: heartbeat(),
+        'heartbeat': lambda **payload: heartbeat,
+        'ping': lambda **payload: ping,
         'stats' : lambda **payload: stats,
         'setprefix': lambda **payload: setprefix,
         'space' : lambda **payload: space,
@@ -32,7 +33,9 @@ commands = {
         'shift' : lambda **payload: shift,
         'status': lambda **payload: status,
         'listener': lambda **payload: listenercmd,
-        'msgstatus': lambda  **payload: msgstatus
+        'msgstatus': lambda  **payload: msgstatus,
+        # Uncomment to load deepfry, requires OpenCV 4
+        # 'deepfry': lambda  **payload: deepfry
     }
 
 @slack.RTMClient.run_on(event='message')
@@ -68,13 +71,14 @@ print(Prefixes.event + 'Custom Plugins Loaded (If Any)')
 try:
     print(Prefixes.event + 'Running Bot...')
     print(Prefixes.start + 'Log Output:')
+    rtmclient.ping_interval = 2
     run_client(rtmclient)
 except KeyboardInterrupt:
     print(Prefixes.event + 'Shutdown Called')
     exit(0)
-# except Exception as e:
-#     bot.error(e)
-#     bot.error_count += 1
-#     print(Prefixes.event + 'Attempting To Auto Reconnect...')
-#     time.sleep(2)
-#     run_client(rtmclient)
+except Exception as e:
+    bot.error(e)
+    bot.error_count += 1
+    print(Prefixes.event + 'Attempting To Auto Reconnect...')
+    time.sleep(2)
+    run_client(rtmclient)